==> data_processing/data_wrangling/7_config_dtypes.ipynb <==
Checking data types in each file and dumping a config-file that can be used to load the data in the correct format.

import pandas as pd
import pickle
import os

dir = "../data/"

dtypes = {
        "id": int, 
        "conversation_id":int,
        "author_id":int,
        "in_reply_to_user_id":str,
        "in_reply_to_username":str,
        "retweeted_user_id":str,
        "retweeted_username":str,
        "quoted_user_id":str,
        "quoted_username":str,
        "created_at":str,
        "lang":str,
        "text":str,
        "public_metrics.retweet_count":float,
        "public_metrics.reply_count":float,
        "public_metrics.retweet_count":float,
        "public_metrics.quote_count":float,
        "public_metrics.like_count":float,
        "public_metrics.impression_count":float,
        "reply_settings":str,
        "possibly_sensitive":str,
        "entities.mentions":str,
        "entities.hashtags":str,
        "entities.annotations":str,
        "entities.cashtags":str,
        "entities.urls":str,
        "referenced_tweets.replied_to.id":str,
        "referenced_tweets.retweeted.id": str, 
        "referenced_tweets.quoted.id": str,
        "author.created_at":str,
        "author.location":str, 
        "author.name":str, 
        "author.username":str, 
        "author.description":str,
        "author.verified":str, 
        "author.verified_type":str,
        "author.protected":str,
        "author.entities.description.cashtags":str,
        "author.entities.description.hashtags":str,
        "author.entities.description.mentions":str,
        "author.entities.description.urls":str,
        "author.public_metrics.followers_count":float, 
        "author.public_metrics.following_count":float,
        "author.public_metrics.tweet_count":float,
        "author.public_metrics.listed_count":float,
        "domain":str,
        "round":str,
        "year":str,
        "month":str,
        "Rating":str,
        "Score":float,
        "Country":str,
        "Language":str,
        "Medium":str,
        "Opinion/Advocacy Journalism":str,
        "Orientation":str,
        "Owner":str,
        "Type of Owner":str,
        "Rating_Date":str,
        "anger_v2":float, 
        "fear_v2":float, 
        "disgust_v2": float, 
        "sadness_v2": float,
        "joy_v2":float, 
        "enthusiasm_v2": float, 
        "pride_v2": float, 
        "hope_v2": float,
        "anger":float, 
        "fear":float, 
        "disgust": float, 
        "sadness": float,
        "joy":float, 
        "enthusiasm": float, 
        "pride": float, 
        "hope": float,
        "type": str,
        "group": str, 
        "not_out": float,
        "out": float
        }

with open(os.path.join(dir,"dtypes_config.pickle"), "wb") as file:
    pickle.dump(dtypes, file)

with open(os.path.join(dir,"dtypes_config.pickle"), "rb") as file:
    dtypes = pickle.load(file)
print(dtypes)
